# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Dantès Finance API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: accounts
    description: Financial accounts
  - name: transactions
    description: Financial transactions
  - name: categories
    description: Transaction categories
  - name: budgets
    description: Budget management
  - name: dashboard
    description: Dashboard summaries

paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /accounts:
    get:
      operationId: listAccounts
      tags: [accounts]
      summary: List all accounts
      responses:
        "200":
          description: List of accounts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Account"
    post:
      operationId: createAccount
      tags: [accounts]
      summary: Create a new account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AccountInput"
      responses:
        "201":
          description: Account created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Account"

  /accounts/{id}:
    get:
      operationId: getAccount
      tags: [accounts]
      summary: Get a specific account
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Account details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Account"
        "404":
          description: Not found
    patch:
      operationId: updateAccount
      tags: [accounts]
      summary: Update an account
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AccountUpdate"
      responses:
        "200":
          description: Updated account
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Account"
    delete:
      operationId: deleteAccount
      tags: [accounts]
      summary: Delete an account
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  /transactions:
    get:
      operationId: listTransactions
      tags: [transactions]
      summary: List all transactions
      parameters:
        - name: accountId
          in: query
          schema:
            type: integer
        - name: categoryId
          in: query
          schema:
            type: integer
        - name: type
          in: query
          schema:
            type: string
            enum: [income, expense, transfer]
        - name: limit
          in: query
          schema:
            type: integer
        - name: offset
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of transactions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Transaction"
    post:
      operationId: createTransaction
      tags: [transactions]
      summary: Create a new transaction
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TransactionInput"
      responses:
        "201":
          description: Transaction created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transaction"

  /transactions/{id}:
    get:
      operationId: getTransaction
      tags: [transactions]
      summary: Get a specific transaction
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Transaction details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transaction"
        "404":
          description: Not found
    patch:
      operationId: updateTransaction
      tags: [transactions]
      summary: Update a transaction
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TransactionUpdate"
      responses:
        "200":
          description: Updated transaction
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transaction"
    delete:
      operationId: deleteTransaction
      tags: [transactions]
      summary: Delete a transaction
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  /categories:
    get:
      operationId: listCategories
      tags: [categories]
      summary: List all categories
      responses:
        "200":
          description: List of categories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Category"
    post:
      operationId: createCategory
      tags: [categories]
      summary: Create a new category
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CategoryInput"
      responses:
        "201":
          description: Category created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"

  /categories/{id}:
    patch:
      operationId: updateCategory
      tags: [categories]
      summary: Update a category
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CategoryUpdate"
      responses:
        "200":
          description: Updated category
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
    delete:
      operationId: deleteCategory
      tags: [categories]
      summary: Delete a category
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  /budgets:
    get:
      operationId: listBudgets
      tags: [budgets]
      summary: List all budgets
      responses:
        "200":
          description: List of budgets
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Budget"
    post:
      operationId: createBudget
      tags: [budgets]
      summary: Create a budget
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BudgetInput"
      responses:
        "201":
          description: Budget created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Budget"

  /budgets/{id}:
    patch:
      operationId: updateBudget
      tags: [budgets]
      summary: Update a budget
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BudgetUpdate"
      responses:
        "200":
          description: Updated budget
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Budget"
    delete:
      operationId: deleteBudget
      tags: [budgets]
      summary: Delete a budget
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  /dashboard/summary:
    get:
      operationId: getDashboardSummary
      tags: [dashboard]
      summary: Get dashboard financial summary
      responses:
        "200":
          description: Dashboard summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardSummary"

  /dashboard/spending-by-category:
    get:
      operationId: getSpendingByCategory
      tags: [dashboard]
      summary: Get spending breakdown by category
      parameters:
        - name: period
          in: query
          schema:
            type: string
            enum: [week, month, quarter, year]
      responses:
        "200":
          description: Spending by category
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SpendingByCategory"

  /dashboard/monthly-flow:
    get:
      operationId: getMonthlyFlow
      tags: [dashboard]
      summary: Get monthly income vs expense flow for the past 12 months
      responses:
        "200":
          description: Monthly flow data
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MonthlyFlow"

  /dashboard/recent-transactions:
    get:
      operationId: getRecentTransactions
      tags: [dashboard]
      summary: Get the most recent transactions across all accounts
      responses:
        "200":
          description: Recent transactions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Transaction"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    Account:
      type: object
      required: [id, name, type, balance, currency, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        type:
          type: string
          enum: [checking, savings, investment, credit, cash]
        balance:
          type: number
        currency:
          type: string
        color:
          type: ["string", "null"]
        icon:
          type: ["string", "null"]
        createdAt:
          type: string

    AccountInput:
      type: object
      required: [name, type, balance, currency]
      properties:
        name:
          type: string
          minLength: 1
        type:
          type: string
          enum: [checking, savings, investment, credit, cash]
        balance:
          type: number
        currency:
          type: string
        color:
          type: string
        icon:
          type: string

    AccountUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        type:
          type: string
          enum: [checking, savings, investment, credit, cash]
        balance:
          type: number
        currency:
          type: string
        color:
          type: string
        icon:
          type: string

    Transaction:
      type: object
      required: [id, accountId, amount, type, description, date, createdAt]
      properties:
        id:
          type: integer
        accountId:
          type: integer
        categoryId:
          type: ["integer", "null"]
        amount:
          type: number
        type:
          type: string
          enum: [income, expense, transfer]
        description:
          type: string
        date:
          type: string
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
        accountName:
          type: ["string", "null"]
        categoryName:
          type: ["string", "null"]
        categoryColor:
          type: ["string", "null"]

    TransactionInput:
      type: object
      required: [accountId, amount, type, description, date]
      properties:
        accountId:
          type: integer
        categoryId:
          type: ["integer", "null"]
        amount:
          type: number
        type:
          type: string
          enum: [income, expense, transfer]
        description:
          type: string
          minLength: 1
        date:
          type: string
        notes:
          type: string

    TransactionUpdate:
      type: object
      properties:
        accountId:
          type: integer
        categoryId:
          type: ["integer", "null"]
        amount:
          type: number
        type:
          type: string
          enum: [income, expense, transfer]
        description:
          type: string
          minLength: 1
        date:
          type: string
        notes:
          type: string

    Category:
      type: object
      required: [id, name, type, color]
      properties:
        id:
          type: integer
        name:
          type: string
        type:
          type: string
          enum: [income, expense, both]
        color:
          type: string
        icon:
          type: ["string", "null"]

    CategoryInput:
      type: object
      required: [name, type, color]
      properties:
        name:
          type: string
          minLength: 1
        type:
          type: string
          enum: [income, expense, both]
        color:
          type: string
        icon:
          type: string

    CategoryUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        type:
          type: string
          enum: [income, expense, both]
        color:
          type: string
        icon:
          type: string

    Budget:
      type: object
      required: [id, categoryId, amount, period, startDate, createdAt]
      properties:
        id:
          type: integer
        categoryId:
          type: integer
        amount:
          type: number
        period:
          type: string
          enum: [weekly, monthly, quarterly, yearly]
        startDate:
          type: string
        endDate:
          type: ["string", "null"]
        createdAt:
          type: string
        categoryName:
          type: ["string", "null"]
        categoryColor:
          type: ["string", "null"]
        spent:
          type: ["number", "null"]

    BudgetInput:
      type: object
      required: [categoryId, amount, period, startDate]
      properties:
        categoryId:
          type: integer
        amount:
          type: number
        period:
          type: string
          enum: [weekly, monthly, quarterly, yearly]
        startDate:
          type: string
        endDate:
          type: string

    BudgetUpdate:
      type: object
      properties:
        categoryId:
          type: integer
        amount:
          type: number
        period:
          type: string
          enum: [weekly, monthly, quarterly, yearly]
        startDate:
          type: string
        endDate:
          type: string

    DashboardSummary:
      type: object
      required: [totalBalance, totalIncome, totalExpenses, netWorth, accountCount, transactionCount]
      properties:
        totalBalance:
          type: number
        totalIncome:
          type: number
        totalExpenses:
          type: number
        netWorth:
          type: number
        accountCount:
          type: integer
        transactionCount:
          type: integer
        savingsRate:
          type: ["number", "null"]

    SpendingByCategory:
      type: object
      required: [categoryId, categoryName, categoryColor, amount, percentage]
      properties:
        categoryId:
          type: integer
        categoryName:
          type: string
        categoryColor:
          type: string
        amount:
          type: number
        percentage:
          type: number

    MonthlyFlow:
      type: object
      required: [month, income, expenses]
      properties:
        month:
          type: string
        income:
          type: number
        expenses:
          type: number
        net:
          type: number