# lib/api-zod/src/generated/types/accountInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Dantès Finance API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AccountInputType } from './accountInputType';

export interface AccountInput {
  /** @minLength 1 */
  name: string;
  type: AccountInputType;
  balance: number;
  currency: string;
  color?: string;
  icon?: string;
}