# lib/api-zod/src/generated/types/categoryUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Dantès Finance API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CategoryUpdateType } from './categoryUpdateType';

export interface CategoryUpdate {
  /** @minLength 1 */
  name?: string;
  type?: CategoryUpdateType;
  color?: string;
  icon?: string;
}