# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Dantès Finance API specification
 * OpenAPI spec version: 0.1.0
 */

export interface DashboardSummary {
  totalBalance: number;
  totalIncome: number;
  totalExpenses: number;
  netWorth: number;
  accountCount: number;
  transactionCount: number;
  /** @nullable */
  savingsRate?: number | null;
}