# lib/api-zod/src/generated/types/transactionUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Dantès Finance API specification
 * OpenAPI spec version: 0.1.0
 */
import type { TransactionUpdateType } from './transactionUpdateType';

export interface TransactionUpdate {
  accountId?: number;
  /** @nullable */
  categoryId?: number | null;
  amount?: number;
  type?: TransactionUpdateType;
  /** @minLength 1 */
  description?: string;
  date?: string;
  notes?: string;
}